fix(appcomposer): refactor Composer commands to avoid build failure
